test_requirements.txt
fabric3
mock
coverage
django-coverage
factory-boy
django-libs
ipdb
flake8
tox
